# docs/swagger.yaml
definitions:
...
  models.Account:
    properties:
      balance:
        type: integer
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      name:
        type: string
      type:
        $ref: '#/definitions/models.AccountType'
      updated_at:
        type: string
    type: object
...
  models.AccountPagedResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/models.Account'
        type: array
      page_number:
        type: integer
      page_size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  models.AccountResponse:
    properties:
      data:
        $ref: '#/definitions/models.Account'
    type: object
  models.AccountType:
    enum:
    - expense
    - income
    - transfer
    type: string
    x-enum-varnames:
    - AccountTypeExpense
    - AccountTypeIncome
    - AccountTypeTransfer
...
  models.CreateAccountRequest:
    properties:
      balance:
        type: integer
      name:
        type: string
      type:
        allOf:
        - $ref: '#/definitions/models.AccountType'
        enum:
        - expense
        - income
        - transfer
    required:
    - name
    - type
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.UpdateAccountRequest:
    properties:
      balance:
        type: integer
      name:
        type: string
      type:
        $ref: '#/definitions/models.AccountType'
    type: object
...
info:
  contact: {}
paths:
  /accounts:
    get:
      consumes:
      - application/json
      description: Get a paginated list of accounts with optional filtering
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - description: Sort by field
        in: query
        name: sort_by
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: sort_order
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AccountPagedResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: List all accounts
      tags:
      - accounts
    post:
      consumes:
      - application/json
      description: Create a new account with name, type and optional balance
      parameters:
      - description: Account creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.AccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Create a new account
      tags:
      - accounts
  /accounts/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete an account by its UUID
      parameters:
      - description: Account UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Delete an account
      tags:
      - accounts
    get:
      consumes:
      - application/json
      description: Get a single account by its UUID
      parameters:
      - description: Account UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get account by ID
      tags:
      - accounts
    put:
      consumes:
      - application/json
      description: Update an existing account by its UUID
      parameters:
      - description: Account UUID
        in: path
        name: id
        required: true
        type: string
      - description: Account update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Update an account
      tags:
      - accounts
...
swagger: "2.0"